+++ openedx/features/learner_profile/views/learner_achievements.py
@@ -7,7 +7,6 @@ from django.template.loader import render_to_string
 from web_fragments.fragment import Fragment
 
 from lms.djangoapps.certificates import api as certificate_api
-from openedx.core.djangoapps.certificates.api import certificates_viewable_for_course
 from openedx.core.djangoapps.content.course_overviews.models import CourseOverview
 from openedx.core.djangoapps.plugin_api.views import EdxFragmentView
 
@@ -45,7 +44,7 @@ class LearnerAchievementsFragmentView(EdxFragmentView):
                 try:
                     course_overview = CourseOverview.get_from_id(course_key)
                     course_certificate['course'] = course_overview
-                    if certificates_viewable_for_course(course_overview):
+                    if certificate_api.certificates_viewable_for_course(course_overview):
                         # add certificate into passing certificate list only if it's a PDF certificate
                         # or there is an active certificate configuration.
                         if course_certificate['is_pdf_certificate'] or course_overview.has_any_active_web_certificate:
